+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Fixes
 
+* Updated the Style of the clear-buttons in the date-time-picker and select component (GUIC-197)
+
 ## 6.4.0 (2018-11-22)
 
 ### Features