+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Fixes
 
+* Remove the firing of the change event in writeValue() method (GUIC-193)
 * Fix the order of ngModel and onChange (GUIC-204)
 
 ## 6.4.0 (2018-11-22)